requirements.txt
Django>=4.2.0
django-cors-headers>=4.0.0
djangorestframework>=3.14.0 
PyJWT>=2.8.0
python-dotenv>=1.0.0
djangorestframework-simplejwt>=5.3.0
mysqlclient>=2.2.0 
google-generativeai>=0.3.0
python-dotenv>=1.0.0
msal